% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{alpha_rubric}
\alias{alpha_rubric}
\title{Cronbach's alpha for a matrix/data frame of rubric components}
\usage{
alpha_rubric(mat)
}
\arguments{
\item{mat}{numeric matrix/data frame with columns = rubric parts}
}
\value{
raw Cronbach's alpha
}
\description{
Cronbach's alpha for a matrix/data frame of rubric components
}
