% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{decisionRubricR-package}
\alias{decisionRubricR}
\alias{decisionRubricR-package}
\title{decisionRubricR: Teaching rubric for human–AI decision quality}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
