% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{skill_corr}
\alias{skill_corr}
\title{Spearman correlation between composite score and a latent trait}
\usage{
skill_corr(comp, latent)
}
\arguments{
\item{comp}{numeric vector of composite scores}

\item{latent}{numeric vector of latent ability/skill}
}
\value{
Spearman's rho
}
\description{
Spearman correlation between composite score and a latent trait
}
